Recompile Babel Relay Plugin
